feat(health): add check endpoint and loop control
On Kubernetes, it makes more sense to use a liveness probe than the
health server loop (i.e. only have one loop). This patch introduces a
flag to disable the health server loop, and a new /check/ endpoint for
such probes.

When the connection is saturated, health checks can take a long time and
therefore fail if the timeout is too short. Gradually increasing the
timeout, as done in the health server loop, is not all that useful
because the upper bound on the timeout is ultimately what you are
willing to tolerate before declaring the connection unhealthy. So a
static probe with a long timeout and a failure count, as implemented in
Kubernetes, will be more stable (i.e. less flopping), especially if more
than one sequential failure is allowed.

The above argument aside, the two health/probe loops also do not work
well together because they can get out of phase. Kubernetes probes
usually must be used to sequence containers in a pod.

Signed-off-by: Jean-Francois Roy <jf@devklog.net>